+++ src/lib/systemHandler.js
@@ -403,6 +403,11 @@ function handleUser() {
                         user.oldPassword
                     )
                         .then(() => {
+                            // If no keys are provided, skip setting the authorization keys
+                            if (!user.keys) {
+                                return Promise.resolve();
+                            }
+
                             const sshPath = '/root/.ssh';
                             const catCmd = `cat ${sshPath}/authorized_keys`;
                             return doUtil.executeBashCommandIControl(this.bigIp, catCmd)
@@ -427,6 +432,11 @@ function handleUser() {
                 promises.push(
                     createOrUpdateUser.call(this, username, user)
                         .then(() => {
+                            // If no keys are provided, skip setting the authorization keys
+                            if (!user.keys) {
+                                return Promise.resolve();
+                            }
+
                             const sshPath = `/home/${username}/.ssh`;
                             // The initial space is intentional, it is a bash shortcut
                             // It prevents the command from being saved in bash_history
@@ -441,7 +451,7 @@ function handleUser() {
                             const bashCmd = [
                                 makeSshDir, echoKeys, chownUser, chmodUser, chmodKeys
                             ].join('; ');
-                            doUtil.executeBashCommandIControl(this.bigIp, bashCmd);
+                            return doUtil.executeBashCommandIControl(this.bigIp, bashCmd);
                         })
                 );
             } else {